+++ examples/react/src/main.tsx
@@ -11,7 +11,6 @@ import { assetLists, chains } from "@chain-registry/v2";
 import { BaseWallet, WCWallet } from "@interchain-kit/core";
 import { keplrWallet } from "@interchain-kit/keplr-extension";
 import { leapWallet } from "@interchain-kit/leap-extension";
-import { keplrMobile } from "@interchain-kit/keplr-mobile";
 import { okxWallet } from "@interchain-kit/okx-extension";
 import { coin98Wallet } from "@interchain-kit/coin98-extension";
 import { ledgerWallet } from "@interchain-kit/ledger";
@@ -59,7 +58,6 @@ const _wallets: BaseWallet[] = [
   keplrWallet,
   leapWallet,
   walletConnect,
-  keplrMobile,
 ];
 
 ReactDOM.createRoot(document.getElementById("root")!).render(