+++ packages/react/src/modal/views/QRCode.tsx
@@ -1,33 +1,37 @@
-import { ConnectModalHead, ConnectModalQRCode } from "@interchain-ui/react"
-import { useActiveWallet, useWalletManager, useWalletModal } from "@interchain-kit/react"
-import { WCWallet } from "@interChain-kit/core"
+import { ConnectModalHead, ConnectModalQRCode } from "@interchain-ui/react";
+import {
+  useActiveWallet,
+  useWalletManager,
+  useWalletModal,
+} from "@interchain-kit/react";
+import { WCWallet } from "@interChain-kit/core";
 
 export const QRCodeHeader = () => {
-  const activeWallet = useActiveWallet()
-  const { close } = useWalletModal()
+  const activeWallet = useActiveWallet();
+  const { close } = useWalletModal();
   return (
     <ConnectModalHead
-      title={activeWallet?.option?.prettyName || ''}
+      title={activeWallet?.option?.prettyName || ""}
       hasBackButton={true}
       onClose={() => void 0}
       onBack={() => void 0}
       closeButtonProps={{ onClick: close }}
     />
-  )
-}
+  );
+};
 export const QRCodeContent = () => {
-  const activeWallet = useActiveWallet() as WCWallet
-  const walletManager = useWalletManager()
-  const data = activeWallet.pairingUri
+  const activeWallet = useActiveWallet() as WCWallet;
+  const walletManager = useWalletManager();
+  const data = activeWallet.pairingUri;
 
   return (
     <ConnectModalQRCode
-      status={data ? "Done" : 'Pending'}
+      status={data ? "Done" : "Pending"}
       link={data}
-      description={'Open App to connect'}
-      errorTitle={'errorTitle'}
-      errorDesc={'errorDesc'}
-      onRefresh={() => walletManager.connect(activeWallet?.option?.name || '')}
+      description={"Open App to connect"}
+      errorTitle={"errorTitle"}
+      errorDesc={activeWallet.errorMessage || ""}
+      onRefresh={() => walletManager.connect(activeWallet?.option?.name || "")}
     />
-  )
-}
+  );
+};