+++ wallets/mock-wallet/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-beta.16](https://github.com/@interchain-kit/mock-wallet/compare/@interchain-kit/mock-wallet@0.0.1-beta.15...@interchain-kit/mock-wallet@0.0.1-beta.16) (2024-09-23)
+
+**Note:** Version bump only for package @interchain-kit/mock-wallet
+
+## [0.0.1-beta.15](https://github.com/@interchain-kit/mock-wallet/compare/@interchain-kit/mock-wallet@0.0.1-beta.14...@interchain-kit/mock-wallet@0.0.1-beta.15) (2024-09-19)
+
+**Note:** Version bump only for package @interchain-kit/mock-wallet
+
+## [0.0.1-beta.14](https://github.com/@interchain-kit/mock-wallet/compare/@interchain-kit/mock-wallet@0.0.1-beta.13...@interchain-kit/mock-wallet@0.0.1-beta.14) (2024-09-19)
+
+**Note:** Version bump only for package @interchain-kit/mock-wallet
+
+## [0.0.1-beta.13](https://github.com/@interchain-kit/mock-wallet/compare/@interchain-kit/mock-wallet@0.0.1-beta.12...@interchain-kit/mock-wallet@0.0.1-beta.13) (2024-09-14)
+
+**Note:** Version bump only for package @interchain-kit/mock-wallet
+
 ## [0.0.1-beta.12](https://github.com/@interchain-kit/mock-wallet/compare/@interchain-kit/mock-wallet@0.0.1-beta.9...@interchain-kit/mock-wallet@0.0.1-beta.12) (2024-09-11)
 
 **Note:** Version bump only for package @interchain-kit/mock-wallet